SUNLIGHT: regulator: mt6358: Add OF match table
The binding documentation mentions that a compatible is required for the
MT6358 device node. But the driver does not provide a OF match table.
This way auto-loading is broken as the MFD driver that registers the
device has a .of_compatible set which makes the platform .uevent
callback report a OF modalias, but that's not in the module.

Bug: N/A
Fixes: f67ff1bd58f0 ("regulator: mt6358: Add support for MT6358 regulator")
Change-Id: I8f4b57a6a1279f8223f6311b2c1d5a6724d9f3ad
Signed-off-by: Matthias Brugger <mbrugger@suse.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>