SUNLIGHT: regulator: mt6323: Add OF match table
The binding documentation mentions that a compatible is required for the
MT6323 device node. But the driver does not provide a OF match table.
This way auto-loading is broken as the MFD driver that registers the
device has a .of_compatible set which makes the platform .uevent
callback report a OF modalias, but that's not in the module.

Bug: N/A
Fixes: 2fdf82923618 ("regulator: mt6323: Add support for MT6323 regulator")
Change-Id: I8e9877e3a066aad6a9e63ed3d0440f30c9bd49b7
Signed-off-by: Matthias Brugger <mbrugger@suse.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>