SUNLIGHT: INTEL_DII: mei: add empty handlers for ops functions
When the offline bit is set and the device is being removed
we should prevent the driver from accessing the hardware because
at this stage the hardware may be no longer available.
Replace the operation handlers with the empty ones to ensure
no hardware registers are being accessed by the driver.

Change-Id: I4dcde678196779be610985f065f920d8a6b5a697
Signed-off-by: Vitaly Lubart <vitaly.lubart@intel.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>