SUNLIGHT: media: dvb-core: Fix use-after-free due to race condition occurring in dvb_frontend
Description:

If the device node of dvb_frontend is open() and the device is
disconnected, many kinds of UAFs may occur when calling close()
on the device node.

The root cause of this is that wake_up() for dvbdev->wait_queue
is implemented in the dvb_frontend_release() function, but
wait_event() is not implemented in the dvb_frontend_stop() function.

So, implement wait_event() function in dvb_frontend_stop() and
add 'remove_mutex' which prevents race condition for 'fe->exit'.

Bug: N/A
Change-Id: I2cb2bd45904dfa6480f3204fe2d347b90031d813
Signed-off-by: Hyunwoo Kim <imv4bel@gmail.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>