SUNLIGHT: ZEN: Add sysctl and CONFIG to disallow unprivileged CLONE_NEWUSER by default
Description:
 - Our default behavior continues to match the vanilla kernel.

Add sysctl and CONFIG to disallow unprivileged CLONE_NEWUSER by default

This is a short-term patch.  Unprivileged use of CLONE_NEWUSER
is certainly an intended feature of user namespaces.  However
for at least saucy we want to make sure that, if any security
issues are found, we have a fail-safe.

Bug: N/A
Change-Id: I4d1e1d7fc5b2e8904a01888d8d09adb2f64ed097
Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>