SUNLIGHT: apparmor4.0.0 [56/87]: prompt - refactor to moving caching to uresponse
BugLink: http://bugs.launchpad.net/bugs/2028253

To be able to be more responsive to prompt return values refactor the
code more to make caching operations distinct from user interaction
and move adding to the cache into the uresps side, so actions don't
have to be passed back to the paused process.

This is a step towards allowing response that will use a named
response instead of permissions and also allow for reuse of knotif
for notifications, heartbeat and canceled.

Change-Id: Id188800b7d5b832c47f3afdca4a542898d293531
Signed-off-by: John Johansen <john.johansen@canonical.com>
(cherry picked from https://gitlab.com/jjohansen/apparmor-kernel)
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>