SUNLIGHT: apparmor4.0.0 [63/87]: prompt - add tailglob on name for cache support
BugLink: http://bugs.launchpad.net/bugs/2028253

Allow the cache to cover more than a single rejection by allowing
an entry to match everything that would match a tailglob expression.
This allows a reduction of cache entries and a reduction of prompts
to userspace when access to a directory hierarchy should be allowed
or denied.

Change-Id: Iee30e104214a3a4ad12159ecd767f5194e91e252
Signed-off-by: John Johansen <john.johansen@canonical.com>
(cherry picked from https://gitlab.com/jjohansen/apparmor-kernel)
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>