SUNLIGHT: apparmor4.0.0 [32/99]: AppArmor: Remove the exclusive flag
BugLink: https://bugs.launchpad.net/bugs/2028253

With the inclusion of the interface LSM process attribute
mechanism AppArmor no longer needs to be treated as an
"exclusive" security module. Remove the flag that indicates
it is exclusive. Remove the stub getpeersec_dgram AppArmor
hook as it has no effect in the single LSM case and
interferes in the multiple LSM case.

Acked-by: Stephen Smalley <stephen.smalley.work@gmail.com>
Acked-by: John Johansen <john.johansen@canonical.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
Change-Id: Iab8ccf15ca8cfd1899bd688e8c9e5953374be9b5
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>