SUNLIGHT: apparmor4.0.0 [21/87]: LSM stacking v39: LSM: security_lsmblob_to_secctx module selection
BugLink: http://bugs.launchpad.net/bugs/2028253

Add a parameter lsmid to security_lsmblob_to_secctx() to identify which
of the security modules that may be active should provide the security
context. If the value of lsmid is LSM_ID_UNDEF the first LSM providing
a hook is used. security_secid_to_secctx() is unchanged, and will
always report the first LSM providing a hook.

Change-Id: I8ef1bffcea5ed4bfa3461136b2f0b28a07a1dccb
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>