SUNLIGHT: kbuild: rpm-pkg: Fix build with non-default MODLIB
Description:

The default MODLIB value is composed of two variables and the hardcoded
string '/lib/modules/'.

MODLIB = $(INSTALL_MOD_PATH)/lib/modules/$(KERNELRELEASE)

Defining this middle part as a variable was rejected on the basis that
users can pass the whole MODLIB to make, such as

make 'MODLIB=$(INSTALL_MOD_PATH)/usr/lib/modules/$(KERNELRELEASE)'

However, this middle part of MODLIB is independently hardcoded by
rpm-pkg, and when the user alters MODLIB this is not reflected when
building the package.

Given that $(INSTALL_MOD_PATH) is overriden during the rpm package build
it is likely going to be empty. Then MODLIB can be passed to the rpm
package, and used in place of the whole
/usr/lib/modules/$(KERNELRELEASE) part.

Change-Id: Ib48ee87b48823b07d257786bcf093e2fc0589b15
Signed-off-by: Michal Suchanek <msuchanek@suse.de>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>