SUNLIGHT: apparmor4.0.0 [70/99]: apparmor: fix notification header size
BugLink: https://bugs.launchpad.net/bugs/2028253

BugLink: http://bugs.launchpad.net/bugs/2028253

When strings are appended to the notification the header size should
be updated to reflect the correct size. While the size is also
directly returned as part of delivering the notification, the header
should also be update to conform to specification and allow for
verification.

If verification is enabled and the notification contains appended
strings then notifications fail verification and won't be delivered.

Fixes: 9a3b87d8b9a0 ("UBUNTU: SAUCE: apparmor4.0.0 [70/99]: prompt - rework build to use append fn, to simplify adding strings")
Change-Id: Ife258d78cb4e397910322a2d2652358e46068e0f
Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>