SUNLIGHT: apparmor4.0.0 [38/87]: LSM stacking v39: LSM: Correct handling of ENOSYS in inode_setxattr
BugLink: http://bugs.launchpad.net/bugs/2028253

The usual "bail on fail" behavior of LSM hooks doesn't
work for security_inode_setxattr(). Modules are allowed
to return -ENOSYS if the attribute specified isn't one
they manage. Fix the code to accommodate this unusal case.
This requires changes to the hooks in SELinux and Smack.

Change-Id: I118a54a56f808f1f85298522e3627bbd1fc52dc5
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>