SUNLIGHT: apparmor4.0.0 [80/87]: apparmor: carry mediation check on label
BugLink: http://bugs.launchpad.net/bugs/2028253

In order to speed up the mediated check, precompute and store the
result as a bit per class type. This not only will allow us to
speed up the mediation check but is also a step to removing the
unconfined special cases as the unconfined check can be replaced
with the generic label_mediates() check.

need to compute initial profile mediated bit vector

Note: label check does not work for capabilities and resources

Change-Id: I025c3e665cc8a1daf868c5ed3b8afbb1873be0aa
Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>