SUNLIGHT: apparmor4.0.0 [06/87]: LSM stacking v39: LSM: Add lsmblob_to_secctx hook
BugLink: http://bugs.launchpad.net/bugs/2028253

Add a new hook security_lsmblob_to_secctx() and its LSM specific
implementations. The LSM specific code will use the lsmblob element
allocated for that module. This allows for the possibility that more
than one module may be called upon to translate a secid to a string,
as can occur in the audit code.

Change-Id: I8b1843e12e5e43fc50235548134661a3464ba660
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>