SUNLIGHT: apparmor4.0.0 [37/87]: LSM stacking v39: LSM: Infrastructure management of the mnt_opts security blob
BugLink: http://bugs.launchpad.net/bugs/2028253

Move management of the mnt_opts->security blob out of the
individual security modules and into the security
infrastructure. Blobs are atill allocated within the modules
as they are only required when mount options are present.
The modules tell the infrastructure how much space is required,
and the space is allocated if needed. Modules can no longer
count on the presence of a blob implying that mount options
specific to that module are present, so flags are added
to the module specific blobs to indicate that this module
has options.

Change-Id: I93156652279dd9519196a0ebeb54b72223224260
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>