SUNLIGHT: apparmor4.0.0 [36/99]: LSM: Infrastructure management of the mnt_opts security blob
BugLink: https://bugs.launchpad.net/bugs/2028253

Move management of the mnt_opts->security blob out of the
individual security modules and into the security
infrastructure. Blobs are atill allocated within the modules
as they are only required when mount options are present.
The modules tell the infrastructure how much space is required,
and the space is allocated if needed. Modules can no longer
count on the presence of a blob implying that mount options
specific to that module are present, so flags are added
to the module specific blobs to indicate that this module
has options.

Change-Id: I33d20bb38bbe9a251d1f3497496bd9010aa068d9
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>