SUNLIGHT: apparmor4.0.0 [84/87]: apparmor: fix x_table_lookup when stacking is not the first entry
BugLink: http://bugs.launchpad.net/bugs/2028253

x_table_lookup currently does stacking during label_parse() if the
target specifies a stack but its only caller ensures that it will
never be used with stacking.

Refactor to slightly simplify the code between in x_to_label(), this
also fixes a long standing problem where x_to_labels check on stacking
is only on the first element to the table option list, instead of
the element that is found and used.

Change-Id: I31154d6490dea1a7a000b80f232f10e3d5151971
Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>