SUNLIGHT: apparmor4.0.0 [70/84]: apparmor: allow profile to be transitioned when a user ns is created
BugLink: http://bugs.launchpad.net/bugs/2028253

User namespaces are used to create sandboxes and often need a different
set of permission than during setup. Allow for a profile state change
on the task creating the namespace.

The transition is encoded the same as exec domain transitions and
only uses a subset of the transitions available at exec. Enivronment
scrubbing etc are not available.

Change-Id: Ifd6863dec06499e02a90a919ec3384d89abcabb8
Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>