SUNLIGHT: apparmor4.0.0 [76/99]: apparmor: ensure labels with more than one entry have correct flags
BugLink: https://bugs.launchpad.net/bugs/2028253

BugLink: http://bugs.launchpad.net/bugs/2028253

labels containing more than one entry need to accumulate flag info
from profiles that the label is constructed from. This is done
correctly for labels created by a merge but is not being done for
labels created by an update or directly created via a parse.

This technically is a bug fix, however the effect in current code is
to cause early unconfined bail out to not happen (ie. its slower) on
labels that were created via update or a parse.

Change-Id: Iafe36338174d7a011915f2be53123e59380e59e7
Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>