platform/x86/intel/pmc/lnl: Remove SSRAM support
A recent PMC firmware change in Lunar Lake caused the pmc_core driver to
fail to probe. This is due to a change in the GUID for PMC telemetry coming
from the SSRAM device. Until a final release is ready this value may
change again. In the meantime, disable the SSRAM support for Lunar Lake so
the driver can load and provide some basic functionality.

Fixes: 3748dfdae2a6 ("platform/x86/intel/pmc: Add Lunar Lake M support to intel_pmc_core driver")
Signed-off-by: David E. Box <david.e.box@linux.intel.com>
Reviewed-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>
Reviewed-by: Ilpo Järvinen <ilpo.jarvinen@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20240227190134.1592072-2-david.e.box@linux.intel.com
Link: https://lore.kernel.org/intel-xe/20240306225350.1910423-1-lucas.demarchi@intel.com/
References: https://gitlab.freedesktop.org/drm/intel/-/issues/10394
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>