SUNLIGHT: apparmor4.0.0 [1/99]: LSM: Infrastructure management of the sock security
BugLink: https://bugs.launchpad.net/bugs/2028253

Move management of the sock->sk_security blob out
of the individual security modules and into the security
infrastructure. Instead of allocating the blobs from within
the modules the modules tell the infrastructure how much
space is required, and the space is allocated there.

Acked-by: Paul Moore <paul@paul-moore.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: John Johansen <john.johansen@canonical.com>
Acked-by: Stephen Smalley <stephen.smalley.work@gmail.com>
Change-Id: I288cd89fdbc34b67de8ac4fae00276093a4dd61d
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>