SUNLIGHT: apparmor4.0.0 [5/99]: LSM: Add lsmblob_to_secctx hook
BugLink: https://bugs.launchpad.net/bugs/2028253

Add a new hook security_lsmblob_to_secctx() and its LSM specific
implementations. The LSM specific code will use the lsmblob element
allocated for that module. This allows for the possibility that more
than one module may be called upon to translate a secid to a string,
as can occur in the audit code.

Change-Id: I94d8f34794532ef371ebac3422185aed8c684b31
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>