SUNLIGHT: apparmor4.0.0 [20/99]: LSM: security_lsmblob_to_secctx module selection
BugLink: https://bugs.launchpad.net/bugs/2028253

Add a parameter lsmid to security_lsmblob_to_secctx() to identify which
of the security modules that may be active should provide the security
context. If the value of lsmid is LSM_ID_UNDEF the first LSM providing
a hook is used. security_secid_to_secctx() is unchanged, and will
always report the first LSM providing a hook.

Change-Id: Ia2e7ae46bbaa7404d4dd026d3bfc01146c7db0f1
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>