SUNLIGHT: Input: psmouse: add NULL check to psmouse_from_serio()
The serio drvdata can be still NULL while the PS/2 interrupt is
processed.  This leaded to crash with a NULL dereference Oops, as
psmouse_from_serio() blindly assumes the non-NULL ps2dev object.

Add a NULL check and return NULL from psmouse_from_serio().  The
returned NULL is handled properly in the caller side, skipping the
rest gracefully.

The log in the bugzilla entry showed that the probe of synaptics
driver succeeded after that point.  So this is a stop-gap solution.

Link: https://bugzilla.suse.com/show_bug.cgi?id=1219522
Change-Id: Id7b2cabe5228b30d0e229498dff200b404a35055
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>