SUNLIGHT: PCI: Mark only devices behind tunneled Ports as "untrusted/removable"
Currently we mark everything behind a PCIe Root Port that has
"ExternalFacingPort" property [1] set as "untrusted" and "removable".
This is fine if the Root Port is tunneled (over Thunderbolt/USB4).

However, in case of a discrete Thunderbolt/USB4 controller it is the
PCIe Downstream Ports of the PCIe switch that are actually tunneled. The
controller typically includes a xHCI (USB 3.x host controller) and NHI
(Thunderbolt/USB4 host controller) as well and with the current logic
these get marked as "untrusted" and "removable" too.

This causes issues if there is a platform policy in place that forbids
attaching drivers against any "untrusted" or "removable" devices. In
addition to this both the xHCI and the NHI get full IOMMU mappings even
though this is not strictly needed, and may degrade their performance
needlessly.

Fix this by marking devices "untrusted" and "removable" only if they are
actually behind a tunneled PCIe Root or Downstream Port. This we can
identify on recent systems using the firmware provided device property,
and for the older ones using the PCI IDs of the known PCIe Root Ports
(this is fixed list, not expected to grow).

[1] https://learn.microsoft.com/en-us/windows-hardware/drivers/pci/dsd-for-pcie-root-ports#identifying-externally-exposed-pcie-root-ports

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Ionut Nechita <ionut_n2001@yahoo.com>