Revert "drm/xe/gsc: define GSC FW for LNL"
This reverts commit bf4d297181f6ce757299a6853620435de76b09d4.

It brought the CI down for LNL. Daniele is investigating the root cause
with the latest update that it has to do with starting the tests in CI
before GSC finishes loading.

Usually commits are simply remove from the topic branch, but this should
be fixed asap with investigation on the way. It's easier to simply
remove the revert when it's time.

Acked-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Link: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/1759
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>